Added 12v boost supply and enlarged board.
This branch and initial commit enlarges the board in order to add a 5V to
12V boost switching supply which can provide up to 400ma of current.

Additionally, J5 was changed to a JST PH connector and J9 and J101 were
added.  All 12V boost supply components start at 100.  Due to the
addition of the boost switching supply, the length of the board was
increased.

In addition, some more filtering capacitors were added and some key traces
for +5V from J9 and J5 were increased to handle additional current for
the 12v supply.

The 12v boost circuit has already been tested and verified to work.

The revision of this branch has been switched to 2.0.0.

Signed-off-by: Aaron Williams <aaron.w2@gmail.com>