2.1.0: Changed NeoPixel to WS2813B-Mini
This replaces the SK6812-3535 with a WS2813B-Mini.  The mini is able to
dissipate heat better.

Also upgraded KiCad to 5.1.2.

Added teardrops to vias.

BOM updated with J5 fixed.

NOTE: The WS2813B-Mini footprint has not yet been tested.  This is a custom
footprint to try and fit the "unique" pattern of the pads.

Signed-off-by: aaronw <aaron.w2@gmail.com>