Fixed microphone, fan PWM and speaker
The ground was not hooked up to the microphone due to an issue with the
footprint.

The PWM pin was also used by the microphone.  Moved PWM pin to IO 5 which is
also shared with 5V output.

The speaker LR clock was using IO34 which is input only.  Moved to IO33 which
is also shared with a 5V output.

Updated revision to 3.0.1.
Fixed page settings, updated dates and revisions.

Signed-off-by: Aaron Williams <aaron.w2@gmail.com>